Start to work on prices.jsp page
I started to work on prices.jsp page, I created a simple page with a table and 3 modal for Create, Update and Delete (only Create is sketchy). I added a datapicker written in jquery and I added a script that orders table's row pressing on the header.

I changed a little bit the method name for PriceDao because of the query with pagination. Not all result are given but only a part, so the method names are not yet retrieveAll* but are simply retrieve*.

I also created a dummy test in order to start to use JUnit.